gein.sh: Don't allow VideoCards='false' & minimal||complete
If a user sets VideoCards to 'false', most likely they don't want a
gui and have selected the improper option. Rather than reaching the
DESKTOP() function and letting emerge error out, prompt the user to
either set VideoCards to allow xorg-drivers to be installed or don't
attempt to install a desktop.