Update variables; improve formatting
* Improve variable naming conventions
* Make all variables lowercase, and NOT camelcase
* Unset variables immediately after they are no longer needed
* Remove most comments and replace with placeholders
* Improve or deprecate some behaviors (ie $VideoCards)

One additional item that I've started work on is using variables to
store the state of whether a "pass" (ie each function) has completed in
order to better control execution, specifically when prior actions are
required. This has not been completed yet though the placeholders have
been added as part of this commit.